Postcopy: Use helpers to map pages during migration
In postcopy, the destination guest is running at the same time
as it's receiving pages; as we receive new pages we must put
them into the guests address space atomically to avoid a running
CPU accessing a partially written page.

Use the helpers in postcopy-ram.c to map these pages.

Note, gcc 4.9.2 is giving me false uninitialized warnings in ram_load's
switch, so anything conditionally set at the start of the switch needs
initializing; filed as
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=64614

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>