Inhibit ballooning during postcopy
The userfault mechanism used for postcopy generates faults
for us on pages that are 'not present', inflating a balloon in
the guest causes host pages to be marked as 'not present'; doing
this during a postcopy, as potentially the same pages were being
received from the source, would confuse the state of the received
page -> disable ballooning during postcopy.

When disabled we drop balloon requests from the guest.  Since ballooning
is generally initiated by the host, the management system should avoid
initiating any balloon instructions to the guest during migration,
although it's not possible to know how long it would take a guest to
process a request made prior to the start of migration.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>