Disallow postcopy on RDMA
The RDMA transport drops RAM contents in using DMA behind the kernel's
back.  This won't cause the userfault for those pages to be disabled.

It might be possible to fix this in the future by changing the RDMA
protocol to have enough information to notify the destination of
incoming pages and bind a separate area as the destination of the RDMA.

Postcopy users should set the capability flag on the destination
to trigger the check.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>