Change state handler/flush
Packets that are sent to the filter queue are usually
sent onwards straight away, but if they can't be
(e.g. the guest is stopped) they sit in the queue until
something causes them to be sent.  Flush the queue
when teh guest restarts.

TODO: I think in a lot of these cases the packets the
queue system stops being sent, should be allowed to be sent when
the guest is stopped.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>